Refactor multer middleware and video route
